Clean up and restructure the codebase
Do some code clean up and restructure the codebase
