--- Frontend/components/FadeInView.js
@@ -1,23 +1,0 @@
-import { Animated, SafeAreaView} from 'react-native';
-import { useRef, useEffect} from 'react';
-export const FadeInView = props => {
-    const fadeAnim = useRef(new Animated.Value(0)).current; // Initial value for opacity: 0
-  
-    useEffect(() => {
-      Animated.timing(fadeAnim, {
-        toValue: 1,
-        duration: 2000,
-        useNativeDriver: true,
-      }).start();
-    }, [fadeAnim]);
-  
-    return (
-      <Animated.View // Special animatable View
-        style={{
-          ...props.style,
-          opacity: fadeAnim, // Bind opacity to animated value
-        }}>
-        {props.children}
-      </Animated.View>
-    );
-  };
